feature: animation implement
detailed:
animation based on event-listener mode.
use of AnimationCurve, create a scriptableobject asset for visualisabe setting

additional:
structure change a little bit - independent factory empty game object for creating and destroying block objects

learned:
use of static - make a class data be class-level, so it can be initialised/edited in other places. So for a class that is dynamically instantiated we can setup its data at the beginning by this way